appcore/requirements.txt
itemadapter==0.6.0
kafka-python==2.0.2
lxml==4.9.1
requests==2.25.1
Scrapy==2.6.1
